refactor: login code from app.vue to presenter and UModal warning fixed (#56)
